+++ app/actions/index.jsx
@@ -1,4 +1,5 @@
 export { default as app } from './app';
 export { default as blockDefs } from './blockDefs';
 export { default as menu } from './menu';
+export { default as message } from './message';
 export { default as user } from './user';