+++ app/components/SideNavigation/SideNavigation.jsx
@@ -1,10 +1,13 @@
+import { MenuList } from '@appsemble/react-bulma';
 import normalize from '@appsemble/utils/normalize';
-import { List } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import React from 'react';
+import { FormattedMessage } from 'react-intl';
+import { NavLink } from 'react-router-dom';
 
-import NavListItem from '../NavListItem';
 import SideMenu from '../SideMenu';
+import messages from './messages';
+import styles from './SideNavigation.css';
 
 /**
  * The app navigation that is displayed in the side menu.
@@ -18,6 +21,13 @@ export default class SideNavigation extends React.Component {
     app: null,
   };
 
+  onLogout = () => {
+    const { closeMenu, logout } = this.props;
+
+    logout();
+    closeMenu();
+  };
+
   render() {
     const { app } = this.props;
 
@@ -28,13 +38,22 @@ export default class SideNavigation extends React.Component {
     return (
       <SideMenu>
         <nav>
-          <List>
+          <MenuList>
             {app.pages.filter(page => !page.parameters).map(page => (
-              <NavListItem key={page.name} to={`/${normalize(page.name)}`}>
-                {page.name}
-              </NavListItem>
+              <li key={page.name}>
+                <NavLink activeClassName="is-active" to={`/${normalize(page.name)}`}>
+                  {page.name}
+                </NavLink>
+              </li>
             ))}
-          </List>
+          </MenuList>
+          <MenuList>
+            <li>
+              <button className={styles.logoutButton} onClick={this.onLogout} type="button">
+                <FormattedMessage {...messages.logout} />
+              </button>
+            </li>
+          </MenuList>
         </nav>
       </SideMenu>
     );