+++ blocks/amsterdam-splash/index.jsx
@@ -25,7 +25,7 @@ attach(({ actions, data }) => {
   }, ANIMATION_LENGTH);
   const root = <div className={styles.root}>{loading}</div>;
   actions.load.dispatch(data).then(
-    () => {
+    response => {
       setTimeout(() => {
         root.replaceChild(
           <header className={styles.content}>
@@ -37,8 +37,10 @@ attach(({ actions, data }) => {
           </header>,
           loading,
         );
+        root.classList.add(styles.done);
+
         setTimeout(() => {
-          actions.success.dispatch(data);
+          actions.success.dispatch(response);
         }, 2e3);
       }, 4e3);
     },