+++ blocks/form/components/GeoCoordinatesInput/GeoCoordinatesInput.jsx
@@ -1,5 +1,5 @@
 import 'leaflet/dist/leaflet.css';
-import { Fas } from '@appsemble/react-bulma';
+import { Button, Icon, Fas } from '@appsemble/react-bulma';
 import PropTypes from 'prop-types';
 import { TileLayer } from 'leaflet/src/layer';
 import { Map } from 'leaflet/src/map';
@@ -22,6 +22,7 @@ export default class GeoCoordinatesInput extends React.Component {
      */
     onChange: PropTypes.func.isRequired,
     reactRoot: PropTypes.instanceOf(HTMLElement).isRequired,
+    utils: PropTypes.shape().isRequired,
   };
 
   locationMarker = new CircleMarker(null, {
@@ -32,9 +33,15 @@ export default class GeoCoordinatesInput extends React.Component {
   ref = React.createRef();
 
   componentDidMount() {
-    const { field, onChange } = this.props;
+    const { field, onChange, utils } = this.props;
 
     const map = new Map(this.ref.current, { attributionControl: false })
+      .once('locationerror', () => {
+        utils.showMessage({
+          // XXX Implement i18n.
+          body: 'Locatie kon niet worden gevonden. Is de locatievoorziening ingeschakeld?',
+        });
+      })
       .on('locationfound', ({ latlng }) => {
         this.locationMarker.setLatLng(latlng).addTo(map);
       })
@@ -55,6 +62,10 @@ export default class GeoCoordinatesInput extends React.Component {
     this.map = map;
   }
 
+  onReset = () => {
+    this.map.setView(this.locationMarker.getLatLng());
+  };
+
   render() {
     return (
       <div className={styles.root}>
@@ -62,6 +73,9 @@ export default class GeoCoordinatesInput extends React.Component {
         <div className={styles.crossHairsOverlay}>
           <Fas className={styles.crossHairs} fa="crosshairs" />
         </div>
+        <Button className={styles.resetButton} onClick={this.onReset}>
+          <Icon className={styles.crossHairs} fa="crosshairs" />
+        </Button>
       </div>
     );
   }